feat: split Settings File
